Support passing AnalyticsClient to SimpleAPI
