+++ README.md
@@ -17,7 +17,7 @@ This project provides 3 ways to interact with the the Coveo usage analytics serv
 
 This JavaScript client project provides a code snippet that a website administrator can easily add to website pages to track page view. This snippet is similar to the Google analytics (analytics.js) snippet.
 
-Note: Currently, the pushed `pageview` events are stored in a Coveo usage analytics table which content currently cannot be viewed in usage analytics reports and the visit browser to prevent performance degradation. But they are used by services such as the Coveo 
+Note: Currently, the pushed `pageview` events are stored in a Coveo usage analytics table which content currently cannot be viewed in usage analytics reports and the visit browser to prevent performance degradation. But they are used by services such as the Coveo
 content recomendation system.
 
 Initially, the `pageview` events data will be used exclusively from Coveo Reveal, a machine learning service (see [Coveo Reveal](http://www.coveo.com/go?dest=cloudhelp&lcid=9&context=177)), by an upcoming feature that will recommend relevant items based on user behavior. It is recommended that you start pushing your events as soon as possible so that you will have data to use when the new feature becomes available.
@@ -27,13 +27,13 @@ Initially, the `pageview` events data will be used exclusively from Coveo Reveal
 1. Get your API key
 
   You need a Coveo Cloud API key that has the permission to write to the usage analytics sevice.
-  
+
   - On [Coveo Cloud V1](https://cloud.coveo.com/), contact [Coveo Support](https://coveocommunity.force.com/) and ask to create an API key with a **Write UA** scope.
 
   - On [Coveo Cloud V2](https://platform.cloud.coveo.com/), create an API key from the [administration console] (https://platform.cloud.coveo.com/admin/#/organization/api-access/) selecting the **Edit** check box for the **Analytics data** privilege (see [API Access - Page](http://www.coveo.com/go?dest=ccv2ac&context=27)).
 
 2. Add the code snippet to all your website pages.
-  
+
   Ask an administrator to add the following code snippet in all pages of your website. Make sure you replace `YOUR_API_KEY` by the API key you got in the previous step.
 
 ```html
@@ -51,7 +51,7 @@ coveoua('send','pageview');
 
 _Note: Do not copy the_ `coveoua.js` _file as it can be updated anytime and you could experience compatibility issues._
 
-#### How to validate events are pushed. 
+#### How to validate events are pushed.
 
   1. In a web browser such as Chrome, navigate to a website page to which you added the code snippet.
   2. In the browser developer tool, go the the **Network** tab.
@@ -100,6 +100,7 @@ npm install
 npm run build:tsc
 npm run build:webpack
 # code code code
+npm run test
 # open pull request
 ```
