+++ src/analyticsclient.ts
@@ -1,0 +1,19 @@
+import {
+    SearchEventRequest, SearchEventResponse,
+    ClickEventRequest, ClickEventResponse,
+    CustomEventRequest, CustomEventResponse,
+    ViewEventRequest, ViewEventResponse,
+    VisitResponse, HealthResponse
+} from "./events";
+
+export interface AnalyticsClient {
+    sendEvent(eventType: string, request: any): Promise<IResponse>;
+    sendSearchEvent(request: SearchEventRequest): Promise<SearchEventResponse>;
+    sendClickEvent(request: ClickEventRequest): Promise<ClickEventResponse>;
+    sendCustomEvent(request: CustomEventRequest): Promise<CustomEventResponse>;
+    sendViewEvent(request: ViewEventRequest): Promise<ViewEventResponse>;
+    getVisit(): Promise<VisitResponse>;
+    getHealth(): Promise<HealthResponse>;
+};
+
+export default AnalyticsClient;
